Create statsd utils
