+++ scripts/workers/submission_worker.py
@@ -26,7 +26,7 @@ from os.path import join
 
 from django.core.files.base import ContentFile
 from django.utils import timezone
-from monitoring.statsd.metrics import NUM_PROCESSED_SUBMISSIONS, increment_statsd_counter
+from statsd_utils import increment_and_push_metrics_to_statsd
 
 # all challenge and submission will be stored in temp directory
 BASE_TEMP_DIR = tempfile.mkdtemp()
@@ -785,21 +785,6 @@ def load_challenge_and_return_max_submissions(q_params):
         challenge.max_concurrent_submission_evaluation
     )
     return maximum_concurrent_submissions, challenge
-
-
-def increment_and_push_metrics_to_statsd(queue_name, is_remote):
-    try:
-        submission_metric_tags = [
-            "queue_name:%s" % queue_name,
-            "is_remote:%d" % is_remote,
-        ]
-        increment_statsd_counter(NUM_PROCESSED_SUBMISSIONS, submission_metric_tags, 1)
-    except Exception as e:
-        logger.exception(
-            "{} Exception when pushing metrics to statsd: {}".format(
-                SUBMISSION_LOGS_PREFIX, e
-            )
-        )
 
 
 def main():
